pyproject.toml
[project]
name = "AskDB"
version = "0.1.0"
description = "AskDB is a Streamlit-powered application that lets you interact with your SQLite databases using natural language queries. It leverages LangChain, LLMs, and Plotly to generate SQL, run queries, and visualize results as interactive charts—all from a simple web interface."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-cohere>=0.4.6",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.8",
    "langchain-ollama>=0.3.7",
    "langchain-sandbox>=0.0.6",
    "langgraph>=0.6.6",
    "langgraph-prebuilt>=0.6.4",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pydantic-ai>=1.0.10",
    "streamlit>=1.49.1",
]
